Fix the uniqueSlice() calculation
Initializing prev to point{} causes its x and y to be nil,
or in this case 0,0. And since 0,0 is a valid point that soon will
be found, the check that prev hasn't been initialized did not work as
intended.

To fix this, prev is now set to the first point in the slice, and
directly added (since it is by definition unique) and then we just
simply go through all points and add those that are not identical
as the previous one.