[WIP] Demo simple test concept
By dividiing the problem in a part1() and part2() function, we can easily
test those with different input to check if they produce the correct output.
This can be useful when re-factoring or touching the code.

It is also possible to easily run those with our specific puzzle input as well
to check that they produce the expected result.

Note: This is just a demo. It currently does not support input options or specifying
which file to read. It will be added later.